Update status label styles and displayed user identifier in the admin and kaprodi views.

+++ database/migrations/2025_09_22_111336_create_konversi2s_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('username');
             $table->foreign('username')->references('username')->on('mahasiswas')->onDelete('cascade');
-            $table->integer('total_sks');
+            $table->float('total_sks');
             $table->enum('status', ['diajukan', 'divalidasi', 'ditolak', 'dikembalikan'])->default('diajukan');
             $table->text('catatan_kaprodi')->nullable();
             $table->timestamps();
